matching-service/requirements.txt
# FastAPI và Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Task Queue
celery==5.3.4
kombu==5.3.4

# Message Broker
pika==1.3.2

# ML/AI Libraries
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
httpx==0.25.2
python-jose[cryptography]==3.3.0

# Monitoring
prometheus-client==0.19.0
